requirements.txt
ib_insync==0.9.86
python-telegram-bot==20.6
python-dotenv==1.0.0
jsonschema==4.19.1
aiosqlite==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
